requirements.txt
streamlit
openai-whisper
torch
numpy
requests